# Conflicts:
#	app/build.gradle.kts
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/uz/androbeck/virtualbank/utils/Constants.kt
#	app/src/main/res/navigation/main_nav_graph.xml
#	app/src/main/res/values-ru/strings.xml
#	app/src/main/res/values-uz/strings.xml
#	app/src/main/res/values/strings.xml
#	gradle/libs.versions.toml

+++ app/src/main/java/uz/androbeck/virtualbank/data/dto/common/response/InComeAndOutComeResDto.kt
@@ -4,15 +4,15 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class InComeAndOutComeResDto (
+data class InComeAndOutComeResDto(
     @SerialName("type")
-    val type: String?=null,
+    val type: String? = null,
     @SerialName("from")
-    val from: String?=null,
+    val from: String? = null,
     @SerialName("to")
-    val to: String?=null,
+    val to: String? = null,
     @SerialName("amount")
-    val amount: Float=0f,
+    val amount: Long = 0,
     @SerialName("time")
-    val time: Long=0
+    val time: Long = 0
 )
